requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
Pillow==10.1.0
stripe==7.8.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
django-debug-toolbar==4.2.0
whitenoise==6.6.0